third_party/blink/web_tests/external/wpt/css/css-pseudo/highlight-pseudos-inheritance-computed-001.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Pseudo-Elements Test: highlight selectors inheritance getComputedStyle</title>
<link rel="help" href="https://drafts.csswg.org/css-pseudo/#highlight-selectors">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #target::selection {
    background-color: green;
    color: lime;
  }
  #target::target-text {
    background-color: green;
    color: lime;
  }
  #target::spelling-error {
    background-color: green;
    color: lime;
  }
  #target::grammar-error {
    background-color: green;
    color: lime;
  }
  #target::highlight(foo) {
    background-color: green;
    color: lime;
  }
</style>
<div id="target"><span id="child"></span></div>
<script>
  for (const pseudo of ["::selection", "::target-text", "::spelling-error", "::grammar-error", "::highlight(foo)"]) {
    test(() => {
      let style = getComputedStyle(child, pseudo);
      assert_equals(style.backgroundColor, "rgb(0, 128, 0)", "Background color is green.");
      assert_equals(style.color, "rgb(0, 255, 0)", "Color is lime.");
    }, `getComputedStyle() for ${pseudo}`);
  }
</script>
